[Component] App Component 만들어 보기

+++ core/commonUtils/src/main/java/com/bowoon/commonutils/RecyclerViewUtils.kt
@@ -2,8 +2,8 @@ package com.bowoon.commonutils
 
 import androidx.recyclerview.widget.RecyclerView
 
-fun scrollPercent(recyclerView: RecyclerView): Double =
-    (recyclerView.computeVerticalScrollOffset() * 1.0 / (recyclerView.computeVerticalScrollRange() - recyclerView.computeVerticalScrollExtent())) * 100.0
+fun scrollPercent(recyclerView: RecyclerView): Float =
+    (recyclerView.computeVerticalScrollOffset() * 1f / (recyclerView.computeVerticalScrollRange() - recyclerView.computeVerticalScrollExtent())) * 100f
 
 //fun RecyclerView.scrollPercent(): Double =
 //    (computeVerticalScrollOffset() * 1.0 / (computeVerticalScrollRange() - computeVerticalScrollExtent())) * 100.0
